+++ app/src/main/java/org/inu/events/ui/home/MainActivity.kt
@@ -16,7 +16,7 @@ import org.inu.events.common.extension.observe
 import org.inu.events.common.extension.observeNonNull
 import org.inu.events.common.extension.toast
 import org.inu.events.databinding.ActivityMainBinding
-import org.inu.events.dialog.LoginDialog
+import org.inu.events.ui.util.dialog.LoginDialog
 import org.inu.events.service.LoginService
 import org.koin.android.ext.android.inject
