+++ app/src/main/java/org/inu/events/ui/mypage/store/LikeViewModel.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
 import org.inu.events.ui.detail.DetailActivity
-import org.inu.events.dialog.TwoButtonDialog
+import org.inu.events.ui.util.dialog.TwoButtonDialog
 import org.inu.events.data.model.dto.LikeParam
 import org.inu.events.data.model.entity.Event
 import org.inu.events.data.repository.LikeRepository
